requirements.txt
streamlit
gspread
oauth2client
google-auth
pandas
streamlit_authenticator